adding conversion MG
